feat: allow scheduling tasks at timestamp
Adds the ability to the stream-platform's ProcessingScheduleService to
run tasks at a timestamp. The execution is not guaranteed to happen
exactly at, but atleast not before the timestamp.

Inline with other task scheduling abilities of the
ProcessingScheduleService, this adds a runAtAsync method as well. It
functions similarly to the other async task scheduling capabilities.

(cherry picked from commit a31d0e42fb8bac4594584347ea657133143cbe1e)