test: generalize more gateway test classes
The test utility classes StubbedBrokerClient and StubbedTopologyManager
move to the common gateway module. They will be reused by the gRPC and REST gateway.

Plus, the GatewayCfgTest moves to the common gateway since it's the unit test for the
configuration class defined in that module.

The gRPC gateway module reuses the test utility classes by depending on the test JAR
of the common gateway module.