feat: use ActivateJobsHandler in REST for long polling
Use an ActivateJobsHandler in the REST API that is instantiated by Spring.
Creates a REST job controller that instructs the handler to look for jobs
to activate. The broker behavior is identical to that over gRPC. Request
cancelation and completion, and collecting intermediate results work a bit
differently due to communication over HTTP 1.1 and servlets. Compared to gRPC,
request cancelation detection and thus deleting pending requests in the broker
is less reliable for REST over HTTP 1.1.