feat: add long-polling handler to REST API
If the REST API is enabled, create a long-polling handler for job activation
that can be used by a job activation endpoint.

The Spring configurations for the broker and gateway modules define the beans required
for the REST controllers separately. Each module can be activated by a specific profile.
Thus, for each profile, the respective beans need to be created.